attempted to study amazons market dominance via introduction of amazon branded products
Outcome
Found our dataset does not support this study.

After attempting to create a dataset of Amazon products and similar products, found that the only candidate Amazon products to study were introduced before or during the beginning of 2018. This means there is not a point of comparison before the introduction of these products.